+++ packages/planum/src/dialog/dialog-close.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react'
+
+import { useDialogState } from './use-dialog-state'
+
+export const DialogClose = React.forwardRef<
+  HTMLButtonElement,
+  React.ButtonHTMLAttributes<HTMLButtonElement>
+>((props, ref) => {
+  const { children, ...rest } = props
+  const state = useDialogState()
+
+  return (
+    <button {...rest} ref={ref} onClick={() => state.setOpen(false)}>
+      {children}
+    </button>
+  )
+})
+
+DialogClose.displayName = 'DialogClose'