+++ packages/planum/src/index.tsx
@@ -66,8 +66,12 @@ export { AutoComplete, AutoCompleteTags } from './auto-complete'
 export type { AutoCompleteProps, AutoCompleteTagsProps } from './auto-complete'
 
 // checkbox
-export { Checkbox } from './checkbox'
-export type { CheckboxProps } from './checkbox'
+export { Checkbox, CheckboxGroup, CheckboxGroupItem } from './checkbox'
+export type {
+  CheckboxProps,
+  CheckboxGroupItemProps,
+  CheckboxGroupProps,
+} from './checkbox'
 
 // Field
 export { Field } from './field'
@@ -127,7 +131,18 @@ export type { LoaderProps } from './loader'
 
 // modal
 export type { ModalProps, ModalState, ModalStateProps } from './modal'
-export { Dialog, Modal, useModalState } from './modal'
+export { Dialog as ModalContent, Modal, useModalState } from './modal'
+
+// dialog
+export {
+  Dialog,
+  DialogClose,
+  DialogPopup,
+  DialogDescription,
+  DialogHeading,
+  DialogTrigger,
+} from './dialog'
+export type { DialogProps } from './dialog'
 
 // separator
 export { Separator } from './separator'