+++ packages/planum/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @uvodohq/planum
 
+## 2.1.0
+
+### Minor Changes
+
+- Checkbox group add
+
+## 2.0.1
+
+### Patch Changes
+
+- Dialog overlay props can be passed
+
+## 2.0.0
+
+### Major Changes
+
+- - Floating v0.11 updated
+  - New Dialog component added
+  - Dialog in Modal renamed to ModalContent
+
 ## 1.1.0
 
 ### Minor Changes